updated user specific routes to check for session

+++ flask_app/flask06.py
@@ -34,8 +34,9 @@ notes = {1: {'title': 'First Note', 'text:': 'This is my first note', 'date':'10
 @app.route('/')
 @app.route('/index')
 def index():
-    a_user =  db.session.query(User).filter_by(email='mogli@uncc.edu').one()
-    return render_template('index.html', user = a_user)
+    if session.get('user'):
+      return render_template("index.html", user=session['user'])
+    return render_template("index.html")
 @app.route('/notes')
 def get_notes():
  if session.get('user'):
@@ -50,22 +51,25 @@ def get_note(note_id):
     return render_template('note.html', note=my_note, user=a_user)
 @app.route('/notes/new', methods = ['GET', 'POST'])
 def new_note():
-    if request.method == 'POST':
+    if session.get('user'):
+      if request.method == 'POST':
        title = request.form['title']
        text = request.form['noteText']
        from datetime import date
        today = date.today()
        today = today.strftime("%m-%d-%Y")
-       new_record = Note(title, text, today)
+       new_record = Note(title, text, today, session['user_id'])
        db.session.add(new_record)
        db.session.commit()
-
-       return redirect(url_for('get_notes'))
+       redirect(url_for('get_notes'))
+      else:
+        return render_template('new.html', user=session['user'])
     else:
-        a_user = db.session.query(User).filter_by(email='mogli@uncc.edu').one()
-        return render_template('new.html', user=a_user)
+        return redirect(url_for('login'))
+
 @app.route('/notes/edit/<note_id>', methods=['GET', 'POST'])
 def update_note(note_id):
+  if session.get('user'):
     if request.method == 'POST':
         title = request.form['title']
         text = request.form['noteText']
@@ -77,17 +81,19 @@ def update_note(note_id):
 
         return redirect(url_for('get_notes'))
     else:
-      a_user = db.session.query(User).filter_by(email='mogli@uncc.edu').one()
       my_note = db.session.query(Note).filter_by(id=note_id).one()
-      return render_template('new.html', note=my_note, user=a_user)
-
+      return render_template('new.html', note=my_note, user= session['user'])
+  else:
+    return redirect(url_for('login'))
 @app.route('/notes/delete/<note_id>', methods=['POST'])
 def delete_note(note_id):
+  if session.get('user'):
     my_note = db.session.query(Note).filter_by(id=note_id).one()
     db.session.delete(my_note)
     db.session.commit()
     return redirect(url_for('get_notes'))
-
+  else:
+    return redirect(url_for('login'))  
 @app.route('/register', methods=['POST', 'GET'])
 def register():
     form = RegisterForm()
